+++ app/src/main/java/org/linphone/core/CoreContext.kt
@@ -96,40 +96,40 @@ class CoreContext
     var isConnectedToAndroidAuto: Boolean = false
 
     val bearerAuthenticationRequestedEvent: MutableLiveData<Event<Pair<String, String?>>> by lazy {
-        MutableLiveData<Event<Pair<String, String?>>>()
+        MutableLiveData()
     }
 
     val digestAuthenticationRequestedEvent: MutableLiveData<Event<String>> by lazy {
-        MutableLiveData<Event<String>>()
+        MutableLiveData()
     }
 
     val clearAuthenticationRequestDialogEvent: MutableLiveData<Event<Boolean>> by lazy {
-        MutableLiveData<Event<Boolean>>()
+        MutableLiveData()
     }
 
     val refreshMicrophoneMuteStateEvent: MutableLiveData<Event<Boolean>> by lazy {
-        MutableLiveData<Event<Boolean>>()
+        MutableLiveData()
     }
 
     val showGreenToastEvent: MutableLiveData<Event<Pair<Int, Int>>> by lazy {
-        MutableLiveData<Event<Pair<Int, Int>>>()
+        MutableLiveData()
     }
 
     val showRedToastEvent: MutableLiveData<Event<Pair<Int, Int>>> by lazy {
-        MutableLiveData<Event<Pair<Int, Int>>>()
+        MutableLiveData()
     }
 
     val showFormattedRedToastEvent: MutableLiveData<Event<Pair<String, Int>>> by lazy {
-        MutableLiveData<Event<Pair<String, Int>>>()
+        MutableLiveData()
     }
 
     val provisioningAppliedEvent: MutableLiveData<Event<Boolean>> by lazy {
-        MutableLiveData<Event<Boolean>>()
+        MutableLiveData()
     }
 
     private var filesToExportToNativeMediaGallery = arrayListOf<String>()
     val filesToExportToNativeMediaGalleryEvent: MutableLiveData<Event<List<String>>> by lazy {
-        MutableLiveData<Event<List<String>>>()
+        MutableLiveData()
     }
 
     private var keepAliveServiceStarted = false
@@ -1040,15 +1040,21 @@ class CoreContext
 
     @WorkerThread
     fun terminateCall(call: Call) {
-        if (call.dir == Call.Dir.Incoming && LinphoneUtils.isCallIncoming(call.state)) {
-            val reason = if (call.core.callsNb > 1) Reason.Busy else Reason.Declined
-            Log.i(
-                "$TAG Declining call [${call.remoteAddress.asStringUriOnly()}] with reason [$reason]"
-            )
-            call.decline(reason)
+        val conference = call.conference
+        if (conference != null) {
+            Log.i("$TAG Terminating conference [${call.remoteAddress.asStringUriOnly()}]")
+            conference.terminate()
         } else {
-            Log.i("$TAG Terminating call [${call.remoteAddress.asStringUriOnly()}]")
-            call.terminate()
+            if (call.dir == Call.Dir.Incoming && LinphoneUtils.isCallIncoming(call.state)) {
+                val reason = if (call.core.callsNb > 1) Reason.Busy else Reason.Declined
+                Log.i(
+                    "$TAG Declining call [${call.remoteAddress.asStringUriOnly()}] with reason [$reason]"
+                )
+                call.decline(reason)
+            } else {
+                Log.i("$TAG Terminating call [${call.remoteAddress.asStringUriOnly()}]")
+                call.terminate()
+            }
         }
     }