+++ app/src/main/java/org/linphone/ui/call/conference/viewmodel/ConferenceViewModel.kt
@@ -83,20 +83,22 @@ class ConferenceViewModel
     val fullScreenMode = MutableLiveData<Boolean>()
 
     val firstParticipantOtherThanOurselvesJoinedEvent: MutableLiveData<Event<Boolean>> by lazy {
-        MutableLiveData<Event<Boolean>>()
+        MutableLiveData()
     }
 
     val showLayoutMenuEvent: MutableLiveData<Event<Boolean>> by lazy {
-        MutableLiveData<Event<Boolean>>()
+        MutableLiveData()
     }
 
     val removeParticipantEvent: MutableLiveData<Event<Pair<String, Participant>>> by lazy {
-        MutableLiveData<Event<Pair<String, Participant>>>()
+        MutableLiveData()
     }
 
     val goToConversationEvent: MutableLiveData<Event<String>> by lazy {
-        MutableLiveData<Event<String>>()
+        MutableLiveData()
     }
+
+    var conferenceConfigured = false
 
     private lateinit var conference: Conference
 
@@ -289,6 +291,7 @@ class ConferenceViewModel
     }
 
     init {
+        conferenceConfigured = false
         isPaused.value = false
         isConversationAvailable.value = false
         isMeParticipantSendingVideo.value = false
@@ -297,6 +300,7 @@ class ConferenceViewModel
 
     @WorkerThread
     fun destroy() {
+        conferenceConfigured = false
         isCurrentCallInConference.postValue(false)
         if (::conference.isInitialized) {
             conference.removeListener(conferenceListener)
@@ -314,6 +318,7 @@ class ConferenceViewModel
         isCurrentCallInConference.postValue(true)
         conference = conf
         conference.addListener(conferenceListener)
+        conferenceConfigured = true
 
         val isIn = conference.isIn
         val state = conf.state
@@ -351,15 +356,15 @@ class ConferenceViewModel
             Log.w(
                 "$TAG Conference has a participant sharing its screen, changing layout from mosaic to active speaker"
             )
-            setNewLayout(ACTIVE_SPEAKER_LAYOUT)
+            setNewLayout(ACTIVE_SPEAKER_LAYOUT, call)
         } else if (currentLayout == AUDIO_ONLY_LAYOUT) {
             val defaultLayout = call.core.defaultConferenceLayout.toInt()
             if (defaultLayout == Conference.Layout.ActiveSpeaker.toInt()) {
                 Log.w("$TAG Joined conference in audio only layout, switching to active speaker layout")
-                setNewLayout(ACTIVE_SPEAKER_LAYOUT)
+                setNewLayout(ACTIVE_SPEAKER_LAYOUT, call)
             } else {
                 Log.w("$TAG Joined conference in audio only layout, switching to grid layout")
-                setNewLayout(GRID_LAYOUT)
+                setNewLayout(GRID_LAYOUT, call)
             }
         }
     }
@@ -461,9 +466,9 @@ class ConferenceViewModel
     }
 
     @WorkerThread
-    fun setNewLayout(newLayout: Int) {
+    fun setNewLayout(newLayout: Int, currentCall: Call? = null) {
         if (::conference.isInitialized) {
-            val call = conference.call
+            val call = conference.call ?: currentCall
             if (call != null) {
                 val params = call.core.createCallParams(call)
                 if (params != null) {